feat: implement BaseField component and refactor field elements to use it
